fix(moe): Ensure dtype consistency when combining expert outputs
The indexed assignment operation (`final_output[mask] = expert_outputs`)
failed during mixed-precision training with `torch.amp.autocast`.

This occurred because `final_output` was maintained as float32, while
`expert_outputs` was cast to float16 by autocast, resulting in a dtype
mismatch.

The fix explicitly casts `expert_outputs` to the dtype of `final_output`
before assignment, making the operation robust across different
environments (CPU, GPU) and precision settings (float16, bfloat16).

+++ model_moe.py
@@ -262,7 +262,7 @@ class MixtureOfExperts(nn.Module):
             # Compute the output from the expert.
             expert_outputs = self.experts[i](expert_inputs)
             # Place the expert outputs into the correct positions in the final output tensor.
-            final_output[token_mask] = expert_outputs
+            final_output[token_mask] = expert_outputs.to(final_output.dtype)
 
         # 5. Scale the output by the gate probability and reshape back.
         # This is a key step to make the routing decision differentiable. The gradient
